Create 15 Days of learning SQL(My SQL)
/*
Julia conducted a 15 days of learning SQL contest. The start date of the contest was March 01, 2016 and the end date was March 15, 2016.

Write a query to print total number of unique hackers who made at least 1 submission each day
(starting on the first day of the contest), and find the hacker_id and name of the hacker who made maximum number 
of submissions each day. If more than one such hacker has a maximum number of submissions, print the lowest hacker_id. 
The query should print this information for each day of the contest, sorted by the date.
*/